pyproject.toml
[project]
name = "nonebot-plugin-mcserver-status"
version = "0.1.0"
description = "An efficient and beautiful Minecraft server status query plugin with vanilla style rendering."
authors = [
    {name = "leiuary", email = "leiuary@qq.com"},
]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0",
    "mcstatus>=11.0.0",
    "Pillow>=10.0.0",
    "pydantic>=1.10.0"
]
requires-python = ">=3.8.1,<4"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "nonebug>=0.4.3; python_version >= '3.9'",
    "nonebot-adapter-onebot>=2.0.0",
]

[project.urls]
"Homepage" = "https://github.com/leiuary/nonebot-plugin-mcserver-status"
"Bug Tracker" = "https://github.com/leiuary/nonebot-plugin-mcserver-status/issues"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_mcserver_status"]
